feat(mcp): instrument local and hosted MCP tool calls
